Fix chapter 3 for knative v0.4.0
Our previous approach for editing the configmap broke because the
autoscaler configmap went from defining the defaults to providing
examples in the configmap on variables to change.

To get around this and keep the changes around this section minimal,
we've added a small configmap to apply that will set this defaults
explicitly.  The remainder of the section (with yq invocations) will
work as advertised.